Refactoring user.entity 패키지 내에서 token, user 패키지 분리
