--- src/main/java/com/themoment/officialgsm/domain/User/entity/Role.java
@@ -1,6 +1,0 @@
-package com.themoment.officialgsm.domain.User.entity;
-
-public enum Role {
-    ADMIN,
-    UNAPPROVED,
-}
