+++ src/main/java/com/themoment/officialgsm/domain/User/entity/user/Role.java
@@ -1,0 +1,6 @@
+package com.themoment.officialgsm.domain.User.entity.user;
+
+public enum Role {
+    ADMIN,
+    UNAPPROVED,
+}
