+++ src/main/java/com/themoment/officialgsm/domain/User/entity/user/User.java
@@ -1,4 +1,4 @@
-package com.themoment.officialgsm.domain.User.entity;
+package com.themoment.officialgsm.domain.User.entity.user;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
