+++ src/main/java/com/themoment/officialgsm/domain/User/repository/BlackListRepository.java
@@ -1,6 +1,6 @@
 package com.themoment.officialgsm.domain.User.repository;
 
-import com.themoment.officialgsm.domain.User.entity.BlackList;
+import com.themoment.officialgsm.domain.User.entity.token.BlackList;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BlackListRepository extends CrudRepository<BlackList, String> {
