+++ src/main/java/com/themoment/officialgsm/domain/User/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
 package com.themoment.officialgsm.domain.User.repository;
 
-import com.themoment.officialgsm.domain.User.entity.RefreshToken;
+import com.themoment.officialgsm.domain.User.entity.token.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
