+++ src/main/java/com/themoment/officialgsm/domain/User/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.themoment.officialgsm.domain.User.repository;
 
-import com.themoment.officialgsm.domain.User.entity.Role;
-import com.themoment.officialgsm.domain.User.entity.User;
+import com.themoment.officialgsm.domain.User.entity.user.Role;
+import com.themoment.officialgsm.domain.User.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
