+++ src/main/java/com/themoment/officialgsm/global/util/UserUtil.java
@@ -1,6 +1,6 @@
 package com.themoment.officialgsm.global.util;
 
-import com.themoment.officialgsm.domain.User.entity.User;
+import com.themoment.officialgsm.domain.User.entity.user.User;
 import com.themoment.officialgsm.domain.User.repository.UserRepository;
 import com.themoment.officialgsm.global.exception.CustomException;
 import lombok.RequiredArgsConstructor;
